order_api/requirements.txt
Flask>=1.1.1
Flask-Cors>=3.0.8
python-dotenv>=0.10.3
elasticsearch==6.4.0
requests>=2.22.0
pytest>=5.1.2
flask-swagger-ui>=3.20.9